pyproject.toml
[project]
name = "zhigalev-travel"
version = "0.1.0"
description = ""
authors = [
    {name = "ILIA",email = "iliazhigalev@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.0,<0.15.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]
